+++ packages/core/src/languages/html/index.ts
@@ -67,7 +67,7 @@ export type HtmlLanguagePluginOptions = {
  *   ],
  * });
  * @since v0.1.0
- * @version v0.1.22
+ * @version v0.1.23
  */
 export default class HtmlLanguagePlugin extends SimpleLanguagePlugin {
   /**