+++ scripts/clean.js
@@ -4,28 +4,69 @@
  * Restore the repository to a clean state, removing all generated files.
  */
 
-import { execFileSync } from "child_process";
+import * as path from "path";
 
+import execSync from "./utilities/exec.js";
+import log from "./utilities/log.js";
 import * as paths from "./paths.js";
 
-const FILES_AND_FOLDERS_TO_DELETE = [
-  ".temp/",
+const HARD_FLAG = "--hard";
+
+const ALWAYS_DELETE = [
   "_reports/",
-  ".eslintcache",
   "npm-debug.log",
-].map(paths.resolve.fromRoot);
+];
 
-const PACKAGES_TO_CLEAN = [
-  "packages/benchmarking",
-  "packages/cli",
-  "packages/core",
-  "packages/testing",
-].map(paths.resolve.fromRoot);
+const HARD_DELETE_ONLY = [
+  ".temp/",
+  ".eslintcache",
+];
 
-execFileSync("git", ["checkout", "HEAD", "--", "./testdata"]);
-execFileSync("rm", ["-rf", ...FILES_AND_FOLDERS_TO_DELETE]);
-PACKAGES_TO_CLEAN.forEach((packageDir) => {
-  execFileSync("npm", ["run", "clean"], { cwd: packageDir });
-});
+main(process.argv);
 
-console.info("Repository cleaned!");
+function main(argv) {
+  argv = argv.slice(2);
+
+  removeFilesAndFolders(argv);
+  resetTestData(argv);
+  cleanPackages(argv);
+}
+
+function removeFilesAndFolders(argv) {
+  log.print("Removing generated files & folders...");
+
+  const filesAndFoldersToRemove = ALWAYS_DELETE;
+  if (argv.includes(HARD_FLAG)) {
+    filesAndFoldersToRemove.push(...HARD_DELETE_ONLY);
+  }
+
+  execSync("rm", [
+    "-rf",
+    ...filesAndFoldersToRemove.map(paths.resolve.fromRoot),
+  ]);
+
+  log.reprintln("Removed generated files & folders.");
+}
+
+function resetTestData() {
+  log.print("Cleaning testdata...");
+  execSync("git", ["checkout", "HEAD", "--", "./testdata"]);
+  log.reprintln("Cleaned testdata.");
+}
+
+function cleanPackages(argv) {
+  if (!argv.includes(HARD_FLAG)) {
+    return;
+  }
+
+  log.println("Cleaning packages...");
+  for (const packageName of paths.getPackages()) {
+    log.print(`  Cleaning packages/${packageName}...`);
+    execSync("npm", ["run", "clean"], {
+      cwd: path.resolve(paths.packagesDir, packageName),
+    });
+    log.reprintln(`  Cleaned packages/${packageName}.`);
+  }
+
+  log.newline();
+}